requirements.txt
configargparse==1.
opencv-python
scipy==1.1.0
tensorboardx
tqdm
pillow==6.2.0
pytorch-lightning==1.5.9
torchvision==0.11.3
scikit_image
setuptools==59.5.0
ruamel.yaml